feat: reorganize into systems and homes
This commit is the first movement towards specifying systems
and homes in a modular fashion so they can be made available
separately.

An open question is whether a specific system, say
systems/bubbles.nix, should define its modules all via
import = [ ./module/name.nix ] or if the mkNixos util should
add some modules.

For now this is how the majority of modules are imported. We
will see how well it works.